refactor(run): move e2e tests to testing directory
