+++ run/events_pubsub/go.mod
@@ -1,5 +1,3 @@
 module github.com/GoogleCloudPlatform/golang-samples/run/events-pubsub
 
 go 1.13
-
-require github.com/GoogleCloudPlatform/golang-samples v0.0.0-20200730192937-48e05c302db2
