+++ run/hello-broken/go.mod
@@ -1,5 +1,3 @@
 module github.com/GoogleCloudPlatform/golang-samples/run/hello-broken
 
 go 1.13
-
-require github.com/GoogleCloudPlatform/golang-samples v0.0.0-20200708192522-c9d21016c9ac
