+++ run/image-processing/go.mod
@@ -5,6 +5,11 @@ go 1.12
 require (
 	cloud.google.com/go v0.60.0
 	cloud.google.com/go/storage v1.10.0
-	github.com/GoogleCloudPlatform/golang-samples v0.0.0-20200709145932-7852be6582ed
+	go.opencensus.io v0.22.4 // indirect
+	golang.org/x/net v0.0.0-20200707034311-ab3426394381 // indirect
+	golang.org/x/text v0.3.3 // indirect
+	golang.org/x/tools v0.0.0-20200708183856-df98bc6d456c // indirect
+	google.golang.org/api v0.29.0 // indirect
 	google.golang.org/genproto v0.0.0-20200708133552-18036109789b
+	google.golang.org/grpc v1.30.0 // indirect
 )
