+++ run/markdown-preview/editor/go.mod
@@ -3,7 +3,12 @@ module github.com/GoogleCloudPlatform/golang-samples/run/markdown-preview/editor
 go 1.13
 
 require (
-	github.com/GoogleCloudPlatform/golang-samples v0.0.0-20200708192522-c9d21016c9ac
+	cloud.google.com/go v0.60.0 // indirect
+	go.opencensus.io v0.22.4 // indirect
+	golang.org/x/net v0.0.0-20200625001655-4c5254603344 // indirect
 	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d
+	golang.org/x/text v0.3.3 // indirect
 	google.golang.org/api v0.29.0
+	google.golang.org/genproto v0.0.0-20200706141556-5779274c8e96 // indirect
+	google.golang.org/grpc v1.30.0 // indirect
 )
