+++ run/markdown-preview/renderer/go.mod
@@ -7,4 +7,5 @@ require (
 	github.com/microcosm-cc/bluemonday v1.0.3
 	github.com/russross/blackfriday/v2 v2.0.1
 	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
+	google.golang.org/api v0.28.0
 )
