+++ run/testing/events_pubsub.e2e_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package testing
 
 import (
 	"fmt"
@@ -24,17 +24,18 @@ import (
 	"github.com/GoogleCloudPlatform/golang-samples/internal/testutil"
 )
 
-func TestBrokenService(t *testing.T) {
+func TestPubSubEventsService(t *testing.T) {
 	tc := testutil.EndToEndTest(t)
 
-	service := cloudrunci.NewService("hello-broken", tc.ProjectID)
+	service := cloudrunci.NewService("run-events-pubsub", tc.ProjectID)
+	service.Dir = "../events_pubsub"
 	if err := service.Deploy(); err != nil {
 		t.Fatalf("service.Deploy %q: %v", service.Name, err)
 	}
 	defer service.Clean()
 
-	requestPath := "/improved"
-	req, err := service.NewRequest("GET", requestPath)
+	requestPath := "/"
+	req, err := service.NewRequest("POST", requestPath)
 	if err != nil {
 		t.Fatalf("service.NewRequest: %v", err)
 	}
@@ -47,7 +48,7 @@ func TestBrokenService(t *testing.T) {
 	defer resp.Body.Close()
 	fmt.Printf("client.Do: %s %s\n", req.Method, req.URL)
 
-	if got := resp.StatusCode; got != http.StatusOK {
-		t.Errorf("response status: got %d, want %d", got, http.StatusOK)
+	if got := resp.StatusCode; got != http.StatusBadRequest {
+		t.Errorf("response status: got %d, want %d", got, http.StatusBadRequest)
 	}
 }
