+++ run/testing/markdown_preview_renderer.e2e_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package testing
 
 import (
 	"fmt"
@@ -24,22 +24,21 @@ import (
 	"github.com/GoogleCloudPlatform/golang-samples/internal/testutil"
 )
 
-func TestDiagramService(t *testing.T) {
+func TestRendererServiceDeploy(t *testing.T) {
 	tc := testutil.EndToEndTest(t)
-	service := cloudrunci.NewService("diagram", tc.ProjectID)
+
+	service := cloudrunci.NewService("renderer", tc.ProjectID)
+	service.Dir = "../markdown-preview/renderer"
 	if err := service.Deploy(); err != nil {
 		t.Fatalf("service.Deploy %q: %v", service.Name, err)
 	}
 	defer service.Clean()
 
-	requestPath := "/diagram.png"
-	req, err := service.NewRequest("GET", requestPath)
+	requestPath := "/"
+	req, err := service.NewRequest("POST", requestPath)
 	if err != nil {
 		t.Fatalf("service.NewRequest: %v", err)
 	}
-	q := req.URL.Query()
-	q.Add("dot", "digraph G { A -> {B, C, D} -> {F} }")
-	req.URL.RawQuery = q.Encode()
 
 	client := http.Client{Timeout: 10 * time.Second}
 	resp, err := client.Do(req)
@@ -49,15 +48,7 @@ func TestDiagramService(t *testing.T) {
 	defer resp.Body.Close()
 	fmt.Printf("client.Do: %s %s\n", req.Method, req.URL)
 
-	if got := resp.StatusCode; got != 200 {
-		t.Errorf("response status: got %d, want %d", got, 200)
-	}
-
-	if got, want := resp.Header.Get("Content-Type"), "image/png"; got != want {
-		t.Errorf("response Content-Type: got %q, want %s", got, want)
-	}
-
-	if got, want := resp.Header.Get("Cache-Control"), "public, max-age=86400"; got != want {
-		t.Errorf("response Cache-Control: got %q, want %q", got, want)
+	if got := resp.StatusCode; got != http.StatusOK {
+		t.Errorf("response status: got %d, want %d", got, http.StatusOK)
 	}
 }
