+++ run/testing/pubsub.e2e_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package testing
 
 import (
 	"fmt"
@@ -28,6 +28,7 @@ func TestPubSubService(t *testing.T) {
 	tc := testutil.EndToEndTest(t)
 
 	service := cloudrunci.NewService("pubsub", tc.ProjectID)
+	service.Dir = "../pubsub"
 	if err := service.Deploy(); err != nil {
 		t.Fatalf("service.Deploy %q: %v", service.Name, err)
 	}
